extract some logics to processor-common
