+++ codeGenerator/common/build.gradle.kts
@@ -3,6 +3,7 @@ plugins {
 }
 group = "org.anime_game_servers.multi_proto"
 version = "0.1"
+
 kotlin {
     jvmToolchain(17)
     jvm {
@@ -16,9 +17,6 @@ kotlin {
         val jvmMain by getting {
             dependencies {
                 implementation("com.google.devtools.ksp:symbol-processing-api:2.1.0-1.0.29")
-                implementation(project(":base"))
-                implementation(project(":processor-common"))
-                implementation("org.anime_game_servers.core:base:0.1")
             }
         }
         val jvmTest by getting
